drivers: sensor: lsm9ds1: fix gyro conversion overflow at 2000 dps
The gyro raw value was multiplied by the sensitivity in 32-bit signed
arithmetic, which overflows at the 2000 dps full scale (70000 udps/LSB),
producing sign-inverted angular rates for large raw values. Perform the
multiplication in 64-bit before narrowing.

+++ drivers/sensor/st/lsm9ds1/lsm9ds1.c
@@ -499,7 +499,7 @@ static inline void lsm9ds1_gyro_convert(struct sensor_value *val, int raw_val, u
 {
 	/* Sensitivity is exposed in udps/LSB */
 	/* Convert to rad/s */
-	sensor_10udegrees_to_rad((raw_val * (int32_t)sensitivity) / 10, val);
+	sensor_10udegrees_to_rad((int32_t)(((int64_t)raw_val * sensitivity) / 10), val);
 }
 
 static inline int lsm9ds1_gyro_get_channel(enum sensor_channel chan, struct sensor_value *val,
